+++ src/http2.h
@@ -160,18 +160,15 @@ nghttp2_nv make_nv(const std::string& name, const std::string& value,
 template<size_t N, size_t M>
 nghttp2_nv make_nv_ll(const char(&name)[N], const char(&value)[M])
 {
-  return { (uint8_t*)name, (uint8_t*)value,
-      (uint16_t)(N - 1), (uint16_t)(M - 1),
-      NGHTTP2_NV_FLAG_NONE };
+  return {(uint8_t*)name, (uint8_t*)value, N - 1, M - 1, NGHTTP2_NV_FLAG_NONE};
 }
 
 // Create nghttp2_nv from string literal |name| and c-string |value|.
 template<size_t N>
 nghttp2_nv make_nv_lc(const char(&name)[N], const char *value)
 {
-  return { (uint8_t*)name, (uint8_t*)value,
-      (uint16_t)(N - 1), (uint16_t)strlen(value),
-      NGHTTP2_NV_FLAG_NONE };
+  return {(uint8_t*)name, (uint8_t*)value, N - 1, strlen(value),
+      NGHTTP2_NV_FLAG_NONE};
 }
 
 // Create nghttp2_nv from string literal |name| and std::string
@@ -179,9 +176,8 @@ nghttp2_nv make_nv_lc(const char(&name)[N], const char *value)
 template<size_t N>
 nghttp2_nv make_nv_ls(const char(&name)[N], const std::string& value)
 {
-  return { (uint8_t*)name, (uint8_t*)value.c_str(),
-      (uint16_t)(N - 1), (uint16_t)value.size(),
-      NGHTTP2_NV_FLAG_NONE };
+  return {(uint8_t*)name, (uint8_t*)value.c_str(), N - 1, value.size(),
+      NGHTTP2_NV_FLAG_NONE};
 }
 
 // Appends headers in |headers| to |nv|. Certain headers, including
